requirements.txt
fastapi>=0.110
uvicorn>=0.27
python-multipart>=0.0.9
azure-cognitiveservices-speech>=1.40
python-dotenv>=1.0
